Fix type casts, constness, parsing and ImGui combo
Apply multiple small fixes and cleanups across debug, tools and misc code:

- Add static_cast conversions (memory usage, M_PI to float in Particle, sf::Mouse positions) to fix type warnings/implicit conversions.
- Make Actor::GetBounds() const and update header/impl accordingly.
- Simplify ImGui combo creation formatting and tighten Render logic to directly invoke the callback when selection changes.
- Harden Animation parsing: properly skip empty/comment lines, add missing braces, improve foundAnimation handling, and add early returns in Update when no frames or when finished.
- Change Array loop indices to unsigned to match size type and avoid signed/unsigned mismatch.
- Remove stray CreateWorld stub/extra brace from SceneManager.hpp.

These changes address compiler warnings, potential UB from type mismatches, and improve parsing/robustness.

+++ GameLoop/Tools/Debug/DebugViewer.cpp
@@ -97,7 +97,7 @@ namespace Debug
 			avgFpsValue = static_cast<int>(totalFps / frameCount);
 			avgFps->SetString("Avg Fps : " + std::to_string(avgFpsValue));
 
-			int memory = GetMemoryUsage();
+			int memory = static_cast<int>(GetMemoryUsage());
 			memoryUsage->SetString("Mem use : " + std::to_string(memory) + "MB");
 		}
 	}
